# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Real-time monitoring for MongoDB: slow queries, index usage, COLLSCAN detection, and current operations.",
        "title": "MongoDB Profiler API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/mongodb-cmd/replica-status": {
            "get": {
                "description": "Gọi rs.status() để lấy thông tin tất cả member: state, health, replication lag. lag_seconds=null nếu member là PRIMARY.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replication"
                ],
                "summary": "Lấy trạng thái Replica Set",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoDB host (phải là member của replica set)",
                        "name": "endpoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/replicaset.ReplicaSetStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "replicaset.MemberStatus": {
            "type": "object",
            "properties": {
                "health": {
                    "description": "1 = up, 0 = down",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "lag_seconds": {
                    "description": "nil nếu là primary",
                    "type": "number"
                },
                "last_error": {
                    "description": "lỗi heartbeat nếu có",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optime_date": {
                    "description": "thời điểm áp dụng oplog gần nhất",
                    "type": "string"
                },
                "self": {
                    "description": "true nếu đây là node đang hỏi",
                    "type": "boolean"
                },
                "state": {
                    "description": "1=PRIMARY, 2=SECONDARY, 7=ARBITER...",
                    "type": "integer"
                },
                "state_str": {
                    "description": "\"PRIMARY\", \"SECONDARY\", \"RECOVERING\"...",
                    "type": "string"
                },
                "uptime": {
                    "description": "giây",
                    "type": "integer"
                }
            }
        },
        "replicaset.ReplicaSetStatus": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "thời điểm lấy status",
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/replicaset.MemberStatus"
                    }
                },
                "my_state": {
                    "description": "state của node đang được hỏi",
                    "type": "integer"
                },
                "set": {
                    "description": "tên replica set",
                    "type": "string"
                }
            }
        },
...
    }
}